requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
gunicorn==21.2.0
SQLAlchemy==1.4.49
asyncpg==0.31.0
alembic==1.11.1
python-dotenv==1.0.0
pydantic==1.10.11
redis==5.0.1
fastapi-limiter==0.1.6
python-multipart==0.0.6
cryptography==41.0.4
